requirements.txt
# Data
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0       # parquet caching

# Analysis
scipy>=1.12.0
statsmodels>=0.14.0
cvxpy>=1.4.0           # portfolio optimization

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.19.0

# Config & utilities
pyyaml>=6.0.1
python-dateutil>=2.9.0
rich>=13.7.0           # pretty terminal output

# Notebooks
jupyter>=1.0.0
ipykernel>=6.29.0
